Refactored
refactored event listeners and merged unnecessary files to declutter.

added mammothWarning.js and associated assets.
